pyproject.toml
[project]
name = "cogent-base"
version = "0.0.1"
description = "Base and shared modules for cogents"
authors = [
    { name = "xmingc", email = "chenxm35@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "litellm>=1.73.0",
    "ollama>=0.4.0",
    "numpy<2.0.0",
    "FlagEmbedding>=1.2.0",
    "unstructured>=0.17.2",
    "filetype>=1.2.0",
    "tomli>=2.0.0",
    "opencv-python>=4.8.0",
    "assemblyai>=0.21.0",
    "weaviate-client>=4.4.0",
    "instructor>=1.7.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "autoflake>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cogent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests as async tests",
]
asyncio_mode = "auto"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 